cmake & compilation speed improvements

+++ Horion/Module/Modules/Reach.h
@@ -6,7 +6,7 @@ private:
 	float reachValue = 3;
 	float originalReach = 0;
 	float* reachPtr = nullptr;
-	DWORD oldProtect = 0;
+	unsigned long oldProtect = 0;
 
 public:
 	Reach();
